Add files via upload
In a file called cash.py in a folder called sentimental-cash, write a program that asks the user how much change is owed and then spits out the minimum number of coins with which said change can be made. 

Use get_float from the CS50 Library to get the user’s input and print to output your answer. Assume that the only coins available are quarters (25¢), dimes (10¢), nickels (5¢), and pennies (1¢).  Write code to calculate change.